refact

+++ src/main/java/tech/dev/investmentms/service/BrapiClient.java
@@ -6,7 +6,6 @@ import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestHeader;
 
-import java.math.BigDecimal;
 import java.util.List;
 
 @FeignClient(
@@ -26,6 +25,6 @@ public interface BrapiClient {
                    String shortName,
                    String longName,
                    String symbol,
-                   BigDecimal regularMarketPrice) {
+                   Double regularMarketPrice) {
     }
 }
